[2nd STAGE] [POC] Move everything from index.native.tsx
